Move TRT utility functions from convert_nodes to convert/utils
